Add condition rules & item card; update idles
Introduce ConditionRule (new Conditions folder) and XPUtility to support unlock conditions and XP logic, add a new Item Card prefab, and update idle data and UI. Wood Idle assets now include startConditions (Wood Idle 2 uses a level requirement) and Wood Idle 1 xpReward was increased. Basic Idle Card prefab and its variant were modified to add icon and lock overlay elements and adjust anchors/sizes; IdleCard and related scripts (IdleData, IdleInstance, IdleManager, Job*, InventoryManager, ShopManager, CurrencyManager, and editor spreadsheet) were updated to accommodate these changes. Also updated Main Scene and added .gitignore entry for ProjectSettings/ProjectVersion.txt.

+++ Assets/Personal/Scripts/ShopManager.cs
@@ -11,7 +11,7 @@ public class ShopManager : MonoBehaviour
     {
         if(inventory.RemoveItem(id, out ItemsData itemData, quantity))
         {
-            currency.AddCurrency(itemData.price * quantity);
+            currency.AddCurrency((ulong)(itemData.price * quantity));
         }
     }
 }
